@@ -1,11 +1,9 @@
-const React      = require('react');
-const ReactDOM   = require('react-dom');
+import React, { useState, useEffect, useContext, useRef} from 'react';
 const $          = require('./sefaria/sefariaJquery');
 const Sefaria    = require('./sefaria/sefaria');
-const classNames = require('classnames');
 const PropTypes  = require('prop-types');
-const Story      = require('./Story');
-
+const { Story }      = require('./Story');
+const { usePaginatedScroll } = require('./Hooks');
 import Component from 'react-class';
 
 // These are duplicated in trend.py needs to be more graceful
@@ -15,179 +13,113 @@ const traits = ["readsHebrew",
 "inDiaspora",
 "inIsrael"];
 
-class StoryEditor extends Component {
-  constructor(props) {
-    super(props);
+const FormFunctions = React.createContext({addStory: null, setError: null});
 
-    this.state = {
-      page: 0,
-      loadedToEnd: false,
-      loading: false,
-      stories: [],
-      submitting: false,
-      error: null
-    };
-  }
-  componentDidMount() {
-    $(ReactDOM.findDOMNode(this)).find(".content").bind("scroll", this.handleScroll);
-    this.getMoreStories();
-  }
-  handleScroll() {
-    if (this.state.loadedToEnd || this.state.loading) { return; }
-    const $scrollable = $(ReactDOM.findDOMNode(this)).find(".content");
-    const margin = 600;
-    if($scrollable.scrollTop() + $scrollable.innerHeight() + margin >= $scrollable[0].scrollHeight) {
-      this.getMoreStories();
-    }
-  }
-  getMoreStories() {
-    $.getJSON("/api/stories?admin_feed=1&page=" + this.state.page, this.loadMoreStories);
-    this.setState({loading: true});
-  }
-  loadMoreStories(data) {
-    if (data.count < data.page_size) {
-      this.setState({loadedToEnd: true});
-    }
-    this.setState({page: data.page + 1, loading: false, stories: this.state.stories.concat(data.stories)});
-  }
-  onDelete(id) {
-    $.ajax({
-        url: '/api/stories/' + id,
-        type: 'DELETE',
-        success: function(result) {
-          if (result.status == "ok") {
-              this.setState({stories: this.state.stories.filter(u => u._id != id)});
-          }
-        }.bind(this)
-    });
-  }
-  removeDraft(timestamp) {
-      this.setState({stories: this.state.stories.filter(u => (!u.draft) || u.timestamp != timestamp)});
-  }
-  addStory(data) {
-      this.state.stories.unshift(data);
-      this.setState({stories: this.state.stories});
-  }
-  handlePublish(type, content, timestamp) {
-    this.setState({"submitting": true, "error": null});
-    const payload = {
-      storyForm: type,
-      data: content
-    };
-    $.ajax({
-      url: "/api/stories",
-      dataType: 'json',
-      type: 'POST',
-      data: {json: JSON.stringify(payload)},
-      success: function(data) {
-        if (data.status == "ok") {
-          payload.date = Date();
-          const stories = this.state.stories.filter(u => (!u.draft) || u.timestamp != timestamp); // Get rid of draft.
-          stories.unshift(payload);
-          this.setState({submitting: false, stories: stories});
+function StoryEditor(props) {
+  const [stories, setStories] = useState([]);
+  const [error, setError] = useState(null);
+  const scrollable_element = useRef();
+
+  usePaginatedScroll(
+      scrollable_element,
+      "/api/stories?admin_feed=1",
+      data => setStories(prev => ([...prev, ...data.stories]))
+  );
+
+  const deleteStory = (id) => $.ajax({url: '/api/stories/' + id, type: 'DELETE',
+    success: (result) => { if (result.status === "ok") { setStories(s => s.filter(u => u._id !== id)) }}
+  });
+  const removeDraft = (timestamp) => setStories(s => s.filter(u => (!u.draft) || u.timestamp !== timestamp));
+  const addStory = (data) => setStories(s => [data, ...s]);
+  const saveStory = (type, content, timestamp) => {
+    setError(null);
+
+    $.ajax({url: "/api/stories", dataType: 'json', type: 'POST',
+      data: {json: JSON.stringify({storyForm: type, data: content})},
+      success: (data) => {
+        if (data.status === "ok") {
+          removeDraft(timestamp);
+          addStory(data.story);
         } else {
-          this.setState({"error": "Error - " + data.error});
+          setError(data.error);
         }
-      }.bind(this),
-      error: function(xhr, status, err) {
-        this.setState({"error": "Error - " + err.toString()});
-        console.error(this.props.url, status, err.toString());
-      }.bind(this)
+      },
+      error: (xhr, status, err) => { setError(err.toString()); }
     });
-  }
-  render() {
-    return (
-      <div className="homeFeedWrapper">
-        <div className="content hasFooter">
-          <div className="contentInner">
-            <h1>
-              <span className="int-en">Stories</span>
-              <span className="int-he">עדכונים</span>
-            </h1>
+  };
 
-            <CreateStoryForm addStory={this.addStory}/>
+  return (<div className="homeFeedWrapper">
+    <div className="content hasFooter" ref={scrollable_element}>
+      <div className="contentInner">
+        <h1>
+          <span className="int-en">Stories</span>
+          <span className="int-he">עדכונים</span>
+        </h1>
 
-            <div className="storyFeed">
-            {this.state.stories.map((s,i) =>
-                [
-                    Story(s,i),
-                    <StoryEditBar
-                        onDelete={this.onDelete}
-                        removeDraft={this.removeDraft}
-                        handlePublish={this.handlePublish}
-                        isDraft={s.draft}
-                        key={s.timestamp + "-" + i + "-editor"}
-                        story={s}/>
-                ]).flat() }
-            </div>
-          </div>
+        <FormFunctions.Provider value={{addStory: addStory, setError: setError}}>
+            <CreateStoryForm/>
+        </FormFunctions.Provider>
+        <span className="error">{error}</span>
+
+        <div className="storyFeed">
+        {stories.map((s,i) =>
+            [
+                Story(s,i, props),
+                <StoryEditBar
+                    deleteStory={deleteStory}
+                    removeDraft={removeDraft}
+                    saveStory={saveStory}
+                    isDraft={s.draft}
+                    key={s.timestamp + "-" + i + "-editor"}
+                    story={s}/>
+            ]).flat() }
         </div>
-      </div>);
-  }
+      </div>
+    </div>
+  </div>);
+
 }
 StoryEditor.propTypes = {
-  interfaceLang:  PropTypes.string
+  interfaceLang:  PropTypes.string,
+  toggleSignUpModal: PropTypes.func
 };
 
-class StoryEditBar extends Component {
-  constructor(props) {
-    super(props);
+function StoryEditBar({story, isDraft, saveStory, deleteStory, removeDraft}) {
+    const [isDeleting, setDeleting] = useState(false);
 
-    this.state = {
-      deleting: false
-    };
-  }
-    handlePublish() {
-        this.props.handlePublish(this.props.story.storyForm, this.props.story.data, this.props.story.timestamp)
+    function onPublish() {
+        saveStory(story.storyForm, story.data, story.timestamp)
     }
-    onDelete() {
-        if(this.props.isDraft) {
-            this.props.removeDraft(this.props.story.timestamp);
+    function onDelete() {
+        if(isDraft) {
+            removeDraft(story.timestamp);
         } else {
-            this.setState({deleting: true});
-            this.props.onDelete(this.props.story._id);
+            setDeleting(true);
+            deleteStory(story._id);
         }
     }
-    render() {
-        return (Sefaria.is_moderator?<div className="storyEditBar">
-            {(this.props.isDraft)?<div className="story-action-button" onClick={this.handlePublish}>Publish</div>:""}
-            {this.state.deleting?<div className="lds-ring"><div></div><div></div><div></div><div></div></div>:
-            <div className="story-action-button" onClick={this.onDelete}>Delete</div>
-            }
-            {this.props.story.mustHave && this.props.story.mustHave.map((trait,i) => <div className="storyEditorTag mustHave" key={i}>{trait}</div>)}
-            {this.props.story.cantHave && this.props.story.cantHave.map((trait,i) => <div className="storyEditorTag cantHave" key={i}>{trait}</div>)}
 
-        </div>:<div/>);
-    }
+    return (Sefaria.is_moderator?<div className="storyEditBar">
+        {(isDraft)?<div className="story-action-button" onClick={onPublish}>Publish</div>:""}
+        {isDeleting?<div className="lds-ring"><div></div><div></div><div></div><div></div></div>:
+                    <div className="story-action-button" onClick={onDelete}>Delete</div>}
+        {story.mustHave && story.mustHave.map((trait,i) => <div className="storyEditorTag mustHave" key={i}>{trait}</div>)}
+        {story.cantHave && story.cantHave.map((trait,i) => <div className="storyEditorTag cantHave" key={i}>{trait}</div>)}
+    </div>:<div/>);
 }
 StoryEditBar.propTypes = {
   interfaceLang:     PropTypes.string,
-  onDelete:          PropTypes.func,
-  handlePublish:     PropTypes.func,
+  deleteStory:       PropTypes.func,
+  saveStory:         PropTypes.func,
   removeDraft:       PropTypes.func,
   isDraft:           PropTypes.bool,
   story:             PropTypes.object
 };
 
+const CreateStoryForm = () => {
+    const [typeName, setTypeName] = useState('New Index');
 
-class CreateStoryForm extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-        typeName: 'New Index',
-        data: {},
-        error: ''
-    };
-  }
-  componentWillReceiveProps(nextProps) {
-    this.setState({"error": nextProps.error});
-  }
-  handleTypeNameChange(e) {
-    this.setState({typeName: e.target.value, error: null});
-  }
-
-  editForms() {
-    return {
+    const editForms = {
         "Free Text":        FreeTextStoryForm,
         "New Index":        NewIndexStoryForm,
         "New Version":      NewVersionStoryForm,
@@ -203,487 +135,331 @@ class CreateStoryForm extends Component {
         groupSheetList: GroupSheetListStory
         */
     };
-  }
 
-  render() {
-      const EditForm = withButton(this.editForms()[this.state.typeName], this.props.addStory);
+    const EditForm = editForms[typeName];
 
-      return (
-          <div className="globalUpdateForm">
-              <div>
-                  <label>
-                      Story Type:
-                      <select value={this.state.typeName} onChange={this.handleTypeNameChange}>
-                          {Object.entries(this.editForms()).map(e => <option value={e[0]} key={e[0]}>{e[0]}</option>)}
-                      </select>
-                  </label>
-              </div>
-              {Sefaria.is_moderator?<EditForm/>:""}
+    return (
+      <div className="globalUpdateForm">
+          <div className="storyTypeSelector">
+              <label>
+                  Story Type:
+                  <select value={typeName} onChange={e => setTypeName(e.target.value)}>
+                      {Object.entries(editForms).map(e => <option value={e[0]} key={e[0]}>{e[0]}</option>)}
+                  </select>
+              </label>
           </div>
-      );
-  }
-}
-CreateStoryForm.propTypes = {
-  addStory:         PropTypes.func
+          {Sefaria.is_moderator?<EditForm/>:""}
+      </div>
+    );
 };
 
-
-function withButton(WrappedFormComponent, addStory) {
-  // ...and returns another component...
-  return class extends React.Component {
-    constructor(props) {
-        super(props);
-        this.state = {
-            submitting: false,
-            error: null,
-        };
-        this.formRef = React.createRef();
-        this.handleReflect = this.handleReflect.bind(this);
-    }
-    handleReflect() {
-        if (!this.formRef.current.isValid()) {
-            this.setState({"error": "Incomplete"});
+function usePreviewButton({payload, isValid}) {
+    const [isSubmitting, setSubmitting] = useState(false);
+    const {addStory, setError} = useContext(FormFunctions);
+    const handleReflect = () => {
+        if (!isValid()) {
+            setError("Incomplete");
             return;
         }
-        this.setState({"submitting": true, "error": null});
+        setSubmitting(true);
+        setError(null);
+        const currentPayload = (payload instanceof Function) ? payload() : payload;
         $.ajax({
             url: "/api/story_reflector",
             dataType: 'json',
             type: 'POST',
-            data: {json: JSON.stringify(this.formRef.current.payload())},
-            success: function (data) {
-                if ("error" in data) {
-                  this.setState({"error": "Error - " + data.error});
-                } else {
+            data: {json: JSON.stringify(currentPayload)},
+            success: (data) => {
+                if ("error" in data) { setError(data.error); }
+                else {
                     data["draft"] = true;
                     addStory(data);
-                    this.setState({submitting: false});
+                    setSubmitting(false);
                 }
-            }.bind(this),
-            error: function (xhr, status, err) {
-                this.setState({"error": "Error - " + err.toString()});
-                console.error(this.props.url, status, err.toString());
-            }.bind(this)
+            },
+            error: (x, s, err) => setError(err.toString())
         });
+    };
+    if (isSubmitting) {
+        return <div className="lds-ring"><div></div><div></div><div></div><div></div></div>;
+    } else {
+        return <input className="previewButton" type="button" value="Preview" onClick={handleReflect}/>
     }
-    render() {
-      // ... and renders the wrapped component with the fresh data!
-      // Notice that we pass through any additional props
-      const disabled = (this.state.submitting ||
-          (this.formRef.current && (!this.formRef.current.isValid())));
-
-      return <div>
-        <WrappedFormComponent ref={this.formRef} {...this.props} />
-        <input type="button" value="Preview" disabled={disabled} onClick={this.handleReflect}/>
-        <span className="error">{this.state.error}</span>
-      </div>;
-    }
-  };
 }
 
 
-class FeaturedSheetsStoryForm extends Component {
-    constructor(props) {
-        super(props);
-        this.state = {
-            error: ''
-        };
-    }
-    payload() {
-        return {
+const FeaturedSheetsStoryForm = () => {
+    const previewButton = usePreviewButton({
+        payload: {
           factory: "SheetListFactory",
-          method: "generate_featured_story",
-        };
-    }
-    isValid() {
-        return true;
-    }
-    render() {
-        return null;
-    }
-}
+          method: "generate_featured_story"
+        },
+        isValid: () => true,
+    });
+    return <div>{previewButton}</div>;
+};
 
-
-class SheetListStoryForm extends Component {
-    constructor(props) {
-        super(props);
-        this.state = {
-            error: ''
-        };
-        this.field_refs = {};
+/* From old is_valid()
+        if (!Object.values(this.field_refs).every(e => e.isValid())) {
+        return false;
     }
-    payload() {
-        const ids_string = this.field_refs.ids.getValue();
-        const ids = ids_string.split(/\s*,\s*/);
-        return {
+    // Required Fields
+    if (!["ids"].every(k => this.field_refs[k].getValue())) {
+        return false;
+    }
+    return true;
+*/
+
+const SheetListStoryForm = () => {
+    const refs = {
+        ids:     useRef(null),
+        lead_en:     useRef(null),
+        lead_he:     useRef(null),
+        title_en:     useRef(null),
+        title_he:     useRef(null)
+    };
+    const previewButton =  usePreviewButton({
+        payload: () => ({
           factory: "SheetListFactory",
           method: "_generate_shared_story",
-          sheet_ids: ids
-        };
-    }
-    isValid() {
-        if (!Object.values(this.field_refs).every(e => e.isValid())) {
-            return false;
-        }
-        // Required Fields
-        if (!["ids"].every(k => this.field_refs[k].getValue())) {
-            return false;
-        }
-        return true;
-    }
-    recordRef(field) {
-        return ref => this.field_refs[field] = ref;
-    }
-    render() {
-        return (
-            <div>
-                <StoryFormTextField placeholder="id, id, id" ref={this.recordRef("ids")} />
-            </div>);
-    }
-}
+          lead: {
+            en:  refs.lead_en.current.getValue(),
+            he:  refs.lead_he.current.getValue()
+          },
+          title: {
+            en: refs.title_en.current.getValue(),
+            he: refs.title_he.current.getValue()
+          },
+          sheet_ids: refs.ids.current.getValue().split(/\s*,\s*/)
+        }),
+        isValid: () => refs.ids.current.isValid() && refs.title_en.current.isValid() && refs.title_he.current.isValid(),
+    });
 
+    return <div>
+        <StoryFormTextField label="Lead (en) (optional)" ref={refs.lead_en} />
+        <StoryFormTextField label="Lead (he) (optional)" ref={refs.lead_he} />
+        <StoryFormTextField label="Title (en)" ref={refs.title_en} />
+        <StoryFormTextField label="Title (he)" ref={refs.title_he} />
+        <StoryFormTextField label="List of IDs" placeholder="id, id, id" ref={refs.ids} />
+        {previewButton}
+    </div>;
+};
 
-
-class UserSheetsStoryForm extends Component {
-    constructor(props) {
-        super(props);
-        this.state = {
-            error: ''
-        };
-        this.field_refs = {};
-    }
-    payload() {
-        return {
+const UserSheetsStoryForm = () => {
+    const refs = {
+        author_uid:     useRef(null),
+    };
+    const previewButton =  usePreviewButton({
+        payload: () => ({
           factory: "UserSheetsFactory",
           method: "_generate_shared_story",
-          author_uid: this.field_refs.author_uid.getValue()
-        };
-    }
-    isValid() {
+          author_uid: refs.author_uid.current.getValue()
+        }),
+        isValid: () => refs.author_uid.current.isValid()
+    });
 
-        if (!Object.values(this.field_refs).every(e => e.isValid())) {
-            return false;
-        }
-        // Required Fields
-        if (!["author_uid"].every(k => this.field_refs[k].getValue())) {
-            return false;
-        }
-        return true;
-    }
-    recordRef(field) {
-        return ref => this.field_refs[field] = ref;
-    }
-    render() {
-        return (
-            <div>
-                <StoryFormTextField placeholder="User ID" ref={this.recordRef("author_uid")} />
-            </div>);
-    }
-}
+    return <div>
+        <StoryFormTextField label="User ID" ref={refs.author_uid} />
+        {previewButton}
+    </div>;
+};
 
-
-class AuthorStoryForm extends Component {
-    constructor(props) {
-        super(props);
-        this.state = {
-            error: ''
-        };
-        this.field_refs = {};
-    }
-    payload() {
-        return {
+const AuthorStoryForm = () => {
+    const refs = {
+        person:     useRef(null),
+    };
+    const previewButton =  usePreviewButton({
+        payload: () => ({
           factory: "AuthorStoryFactory",
           method: "_generate_shared_story",
-          person: this.field_refs.person.getValue()
-        };
-    }
-    isValid() {
+          person: refs.person.current.getValue()
+        }),
+        isValid: () => refs.person.current.isValid()
+    });
 
-        if (!Object.values(this.field_refs).every(e => e.isValid())) {
-            return false;
-        }
-        // Required Fields
-        if (!["person"].every(k => this.field_refs[k].getValue())) {
-            return false;
-        }
-        return true;
-    }
-    recordRef(field) {
-        return ref => this.field_refs[field] = ref;
-    }
-    render() {
-        return (
-            <div>
-                <StoryFormTextField placeholder="Author Key" ref={this.recordRef("person")} />
-            </div>);
-    }
-}
+    return <div>
+        <StoryFormTextField label="Author Key" ref={refs.person} />
+        {previewButton}
+    </div>;
+};
 
-
-class TopicSourcesStoryForm extends Component {
-    constructor(props) {
-        super(props);
-        this.state = {
-            error: ''
-        };
-        this.field_refs = {};
-    }
-    payload() {
-        return {
+const TopicSourcesStoryForm = () => {
+    const refs = {
+        topic:     useRef(null),
+    };
+    const previewButton =  usePreviewButton({
+        payload: () => ({
           factory: "TopicTextsStoryFactory",
           method: "_generate_shared_story",
-          topic: this.field_refs.topic.getValue()
-        };
-    }
-    isValid() {
+          topic: refs.topic.current.getValue()
+        }),
+        isValid: () => refs.topic.current.isValid()
+    });
 
-        if (!Object.values(this.field_refs).every(e => e.isValid())) {
-            return false;
-        }
-        // Required Fields
-        if (!["topic"].every(k => this.field_refs[k].getValue())) {
-            return false;
-        }
-        return true;
-    }
-    recordRef(field) {
-        return ref => this.field_refs[field] = ref;
-    }
-    render() {
-        return (
-            <div>
-                <StoryFormTextField placeholder="Topic" ref={this.recordRef("topic")} />
-            </div>);
-    }
-}
+    return <div>
+        <StoryFormTextField label="Topic" ref={refs.topic} />
+        {previewButton}
+    </div>;
+};
 
-class TopicSheetsStoryForm extends Component {
-    constructor(props) {
-        super(props);
-        this.state = {
-            error: ''
-        };
-        this.field_refs = {};
-    }
-    payload() {
-        return {
+const TopicSheetsStoryForm = () => {
+    const refs = {
+        topic:     useRef(null),
+    };
+    const previewButton =  usePreviewButton({
+        payload: () => ({
           factory: "SheetListFactory",
           method: "generate_topic_story",
-          topic: this.field_refs.topic.getValue()
-        };
-    }
-    isValid() {
+          topic: refs.topic.current.getValue()
+        }),
+        isValid: () => refs.topic.current.isValid()
+    });
 
-        if (!Object.values(this.field_refs).every(e => e.isValid())) {
-            return false;
-        }
-        // Required Fields
-        if (!["topic"].every(k => this.field_refs[k].getValue())) {
-            return false;
-        }
-        return true;
-    }
-    recordRef(field) {
-        return ref => this.field_refs[field] = ref;
-    }
-    render() {
-        return (
-            <div>
-                <StoryFormTextField placeholder="Topic" ref={this.recordRef("topic")} />
-            </div>);
-    }
-}
-
-class TextPassageStoryForm extends Component {
-    constructor(props) {
-        super(props);
-        this.state = {
-            type: 'textPassage',
-            error: ''
-        };
-        this.field_refs = {};
-    }
-    payload() {
-        const d = { ref: this.field_refs.ref.getValue() };
-        return {
-          storyForm: this.state.type,
-          data: d
-        };
-    }
-    isValid() {
-
-        if (!Object.values(this.field_refs).every(e => e.isValid())) {
-            return false;
-        }
-        // Required Fields
-        if (!["ref"].every(k => this.field_refs[k].getValue())) {
-            return false;
-        }
-        return true;
-    }
-    recordRef(field) {
-        return ref => this.field_refs[field] = ref;
-    }
-    render() {
-        return (
-            <div>
-                <StoryFormRefField ref={this.recordRef("ref")} />
-            </div>);
-    }
-}
-TextPassageStoryForm.propTypes = {
-    
+    return <div>
+        <StoryFormTextField label="Topic" ref={refs.topic} />
+        {previewButton}
+    </div>;
 };
 
-class FreeTextStoryForm extends Component {
-    constructor(props) {
-        super(props);
-        this.state = {
-            type: 'freeText',
-            error: ''
-        };
-        this.field_refs = {};
-    }
-    payload() {
-        const d = {
-            en: this.field_refs.en.getValue(),
-            he: this.field_refs.he.getValue(),
-        };
-        return {
-          storyForm: this.state.type,
-          data: d
-        };
-    }
-    isValid() {
-        if (!Object.values(this.field_refs).every(e => e.isValid())) {
-            return false;
-        }
-        // Required Fields
-        if (!["en","he"].every(k => this.field_refs[k].getValue())) {
-            return false;
-        }
-        return true;
-    }
-    recordRef(field) {
-        return ref => this.field_refs[field] = ref;
-    }
-    render() {
-        return (
-            <div>
-                <StoryFormTextareaField ref={this.recordRef("en")} placeholder="English"/>
-                <StoryFormTextareaField ref={this.recordRef("he")} placeholder="Hebrew"/>
-            </div>);
-    }
-}
-FreeTextStoryForm.propTypes = {
-    
+const TextPassageStoryForm = () => {
+    const refs = {
+        ref:     useRef(null),
+        lead_en:     useRef(null),
+        lead_he:     useRef(null),
+        title_en:     useRef(null),
+        title_he:     useRef(null)
+    };
+    const previewButton =  usePreviewButton({
+        payload: () => ({
+            storyForm: 'textPassage',
+            data: {
+                lead: {
+                    en:  refs.lead_en.current.getValue(),
+                    he:  refs.lead_he.current.getValue()
+                },
+                title: {
+                    en: refs.title_en.current.getValue(),
+                    he: refs.title_he.current.getValue()
+                },
+                ref: refs.ref.current.getValue(),
+            }
+        }),
+        isValid: () => refs.ref.current.isValid() && refs.title_en.current.isValid() && refs.title_he.current.isValid()
+    });
+
+    return <div>
+        <StoryFormTextField label="Lead (en) (optional)" ref={refs.lead_en} />
+        <StoryFormTextField label="Lead (he) (optional)" ref={refs.lead_he} />
+        <StoryFormTextField label="Title (en)" ref={refs.title_en} />
+        <StoryFormTextField label="Title (he)" ref={refs.title_he} />
+        <StoryFormRefField label="Ref" ref={refs.ref} />
+
+        {previewButton}
+    </div>;
 };
 
-class NewVersionStoryForm extends Component {
-    constructor(props) {
-        super(props);
-        this.state = {
-            type: 'newVersion',
-            error: ''
-        };
-        this.field_refs = {};
-    }
-    payload() {
-        const d = { index: this.field_refs.index.getValue() };
-        Object.assign(d, this.field_refs.version.getValue());
-        ["ref","en","he"].forEach(p => {if (!!this.field_refs[p].getValue()) {d[p] = this.field_refs[p].getValue()}});
-        return {
-          storyForm: this.state.type,
-          data: d
-        };
-    }
-    isValid() {
-        if (!Object.values(this.field_refs).every(e => e.isValid())) {
-            return false;
-        }
-        // Required Fields
-        if (!["index"].every(k => this.field_refs[k].getValue())) {
-            return false;
-        }
-        return true;
-    }
-    recordRef(field) {
-        return ref => this.field_refs[field] = ref;
-    }
-    render() {
-        return (
-            <div>
-                <StoryFormIndexField ref={this.recordRef("index")}/>
-                <StoryFormVersionFields ref={this.recordRef("version")}/>
-                <StoryFormTextareaField ref={this.recordRef("en")} placeholder="English Description (optional)"/>
-                <StoryFormTextareaField ref={this.recordRef("he")} placeholder="Hebrew Description (optional)"/>
-                <StoryFormRefField ref={this.recordRef("ref")} />
-            </div>);
-    }
-}
-NewVersionStoryForm.propTypes = {
-    
+const FreeTextStoryForm = () => {
+    const refs = {
+        en:     useRef(null),
+        he:     useRef(null),
+    };
+    const previewButton =  usePreviewButton({
+        payload: () => {
+            const d = Object.keys(refs).reduce((obj, k) => {
+                    const v = refs[k].current.getValue();
+                    if (v) { obj[k] = v; }
+                    return obj;
+                }, {});
+            return {
+                storyForm: 'freeText',
+                data: d
+                }
+            },
+        isValid: () => (refs.en.current.isValid() && refs.he.current.isValid())
+    });
+
+    return <div>
+        <StoryFormTextareaField ref={refs.en} label="English"/>
+        <StoryFormTextareaField ref={refs.he} label="Hebrew"/>
+        {previewButton}
+    </div>;
 };
 
-class NewIndexStoryForm extends Component {
-    constructor(props) {
-        super(props);
-        this.state = {
-            type: 'newIndex',
-            error: ''
-        };
-        this.field_refs = {};
-    }
-    payload() {
-        const d = { index: this.field_refs.index.getValue() };
-        ["ref","en","he"].forEach(p => {if (!!this.field_refs[p].getValue()) {d[p] = this.field_refs[p].getValue()}});
-        return {
-          storyForm: this.state.type,
-          data: d
-        };
-    }
-    isValid() {
-/*
-        if (!Object.values(this.field_refs).every(e => e.isValid())) {
-            return false;
-        }
-        // Required Fields
-        if (!["index"].every(k => this.field_refs[k].getValue())) {
-            return false;
-        } */
-        return true;
-    }
-    recordRef(field) {
-        return ref => this.field_refs[field] = ref;
-    }
-    render() {
-        return (
-            <div>
-                <StoryFormIndexField ref={this.recordRef("index")}/>
-                <StoryFormTextareaField ref={this.recordRef("en")} placeholder="English Description (optional)"/>
-                <StoryFormTextareaField ref={this.recordRef("he")} placeholder="Hebrew Description (optional)"/>
-                <StoryFormRefField ref={this.recordRef("ref")} />
-            </div>);
-    }
-}
-NewIndexStoryForm.propTypes = {
-    
+const NewVersionStoryForm = () => {
+    const refs = {
+        index:  useRef(null),
+        version: useRef(null),
+        en:     useRef(null),
+        he:     useRef(null),
+        ref:    useRef(null),
+    };
+    const previewButton =  usePreviewButton({
+        payload: () => {
+            const d = Object.keys(refs).reduce((obj, k) => {
+                    const v = refs[k].current.getValue();
+                    if (v) { obj[k] = v; }
+                    return obj;
+                }, {});
+            return {
+                storyForm: 'newVersion',
+                data: d
+                }
+            },
+        isValid: () => (refs.index.current.isValid() && refs.version.current.isValid()),
+    });
+
+    return <div>
+        <StoryFormIndexField ref={refs.index}/>
+        <StoryFormVersionFields ref={refs.version}/>
+        <StoryFormTextareaField ref={refs.en} label="English Description (optional)"/>
+        <StoryFormTextareaField ref={refs.he} label="Hebrew Description (optional)"/>
+        <StoryFormRefField ref={refs.ref} label="Ref (optional)" />
+        {previewButton}
+    </div>;
+};
+
+const NewIndexStoryForm = () => {
+    const refs = {
+        index:  useRef(null),
+        en:     useRef(null),
+        he:     useRef(null),
+        ref:    useRef(null),
+    };
+    const previewButton =  usePreviewButton({
+        payload: () => {
+            const d = Object.keys(refs).reduce((obj, k) => {
+                    const v = refs[k].current.getValue();
+                    if (v) { obj[k] = v; }
+                    return obj;
+                }, {});
+            return {
+                storyForm: 'newIndex',
+                data: d
+                }
+            },
+        isValid: () => refs.index.current.isValid()
+    });
+
+    return <div>
+        <StoryFormIndexField ref={refs.index} />
+        <StoryFormTextareaField ref={refs.en} label="English Description (optional)" />
+        <StoryFormTextareaField ref={refs.he} label="Hebrew Description (optional)" />
+        <StoryFormRefField ref={refs.ref} label="Ref (optional)" />
+        {previewButton}
+    </div>;
 };
 
 
-
+/********************************/
 /*          Fields              */
-
+/********************************/
 
 
 class StoryFormIndexField extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-        value: '',
-        error: ''
-    };
-  }
+    constructor(props) {
+        super(props);
+        this.state = {value: '', error: ''};
+    }
     getValue() {
         return this.state.value.trim();
     }
@@ -691,35 +467,25 @@ class StoryFormIndexField extends Component {
         return (this.state.value && (!this.state.error));
     }
     handleChange(e) {
-        this.setState({value: e.target.value});
-        this.validate();
-    }
-    validate() {
-      /*
-        if (Sefaria.index(this.state.value)) {
-            this.setState({error: ''});
-        } else {
-            this.setState({error: "Not an Index"});
-        }
-        */
+        const value = e.target.value;
+        const error = (Sefaria.index(value)) ? '' : "Not an Index";
+        this.setState({value: value, error: error});
     }
     render() {
       return <div>
-        <input type="text" placeholder="Index Title" onChange={this.handleChange} />
+        <label>Index
+            <input type="text" onChange={this.handleChange} />
+        </label>
         <span className="error">{this.state.error}</span>
       </div>;
   }
 }
 
 class StoryFormVersionFields extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-        language: "en",
-        version: '',
-        error: ''
-    };
-  }
+    constructor(props) {
+        super(props);
+        this.state = {value: '', error: '', language: "en"};
+    }
     getValue() {
         return {
             language: this.state.language,
@@ -731,35 +497,33 @@ class StoryFormVersionFields extends Component {
     }
     handleLanguageChange(e) {
         this.setState({language: e.target.value});
-        this.validate();
+        // Validate
     }
     handleVersionChange(e) {
         this.setState({version: e.target.value});
-        this.validate();
-    }
-    validate() {
         //todo: See if this is a valid version
     }
     render() {
       return <div>
-        <select type="text" value={this.state.language} placeholder="Version Language" onChange={this.handleLanguageChange}>
-            <option value="en">English</option>
-            <option value="he">Hebrew</option>
-          </select>
-        <input type="text" value={this.state.version} placeholder="Version Title" onChange={this.handleVersionChange}/>
+        <label>Version Language
+            <select type="text" value={this.state.language} onChange={this.handleLanguageChange}>
+                <option value="en">English</option>
+                <option value="he">Hebrew</option>
+            </select>
+        </label>
+        <label>Version Title
+            <input type="text" value={this.state.version} onChange={this.handleVersionChange}/>
+        </label>
         <span className="error">{this.state.error}</span>
       </div>
     }
 }
 
 class StoryFormTextareaField extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-        value: '',
-        error: ''
-    };
-  }
+    constructor(props) {
+        super(props);
+        this.state = {value: '', error: ''};
+    }
     getValue() {
         return this.state.value.trim();
     }
@@ -768,32 +532,32 @@ class StoryFormTextareaField extends Component {
     }
     handleChange(e) {
         this.setState({value: e.target.value});
-        this.validate();
-    }
-    validate() {
-
+        // Validate
     }
     render() {
       return <div>
-        <textarea
-            placeholder={this.props.placeholder}
-            value={this.state.value}
-            onChange={this.handleChange}
-            rows="3"
-            cols="80" />
+          <label>{this.props.label}
+            <textarea
+                placeholder={this.props.placeholder}
+                value={this.state.value}
+                onChange={this.handleChange}
+                rows="3"
+                cols="80" />
+          </label>
         <span className="error">{this.state.error}</span>
       </div>;
     }
 }
+StoryFormTextareaField.propTypes = {
+    placeholder:    PropTypes.string,
+    label:          PropTypes.string,
+};
 
 class StoryFormTextField extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-        value: '',
-        error: ''
-    };
-  }
+    constructor(props) {
+        super(props);
+        this.state = {value: '', error: ''};
+    }
     getValue() {
         return this.state.value.trim();
     }
@@ -802,29 +566,26 @@ class StoryFormTextField extends Component {
     }
     handleChange(e) {
         this.setState({value: e.target.value});
-        this.validate();
-    }
-    validate() {
-
+        // Validate
     }
     render() {
       return <div>
-            <input type="text" value={this.state.value} placeholder={this.props.placeholder} onChange={this.handleChange}/>
+            <label>{this.props.label}
+                <input type="text" value={this.state.value} placeholder={this.props.placeholder} onChange={this.handleChange}/>
+            </label>
             <span className="error">{this.state.error}</span>
       </div>;
     }
 }
 StoryFormTextField.propTypes = {
-    placeholder: PropTypes.string
+    placeholder:    PropTypes.string,
+    label:          PropTypes.string,
 };
 
 class StoryFormRefField extends Component {
     constructor(props) {
         super(props);
-        this.state = {
-            value: '',
-            error: ''
-        };
+        this.state = {value: '', error: ''};
     }
 
     getValue() {
@@ -836,24 +597,24 @@ class StoryFormRefField extends Component {
     }
 
     handleChange(e) {
-        this.setState({value: e.target.value});
-        this.validate();
-    }
-    validate() {
-        if (Sefaria.isRef(this.state.value)) {
-            this.setState({error: ''});
-        } else {
-            this.setState({error: "Not a Ref"});
-        }
+        const value = e.target.value;
+        const error = (Sefaria.isRef(value)) ? '' : "Not a Ref";
+        this.setState({value: value, error: error});
     }
 
     render() {
         return <div>
-            <input type="text" value={this.state.value} placeholder="Ref" onChange={this.handleChange}/>
+            <label>{this.props.label}
+                <input type="text" value={this.state.value} onChange={this.handleChange}/>
+            </label>
             <span className="error">{this.state.error}</span>
         </div>;
     }
 }
+StoryFormRefField.propTypes = {
+    label: PropTypes.string,
+};
+
 
 class StoryFormUserField extends Component {
 
