@@ -15,7 +15,7 @@ import regex
 from . import abstract as abst
 from sefaria.system.database import db
 from sefaria.model.lexicon import LexiconEntrySet
-from sefaria.system.exceptions import InputError, IndexSchemaError
+from sefaria.system.exceptions import InputError, IndexSchemaError, DictionaryEntryNotFoundError, SheetNotFoundError
 from sefaria.utils.hebrew import decode_hebrew_numeral, encode_small_hebrew_numeral, encode_hebrew_numeral, encode_hebrew_daf, hebrew_term, sanitize
 
 """
@@ -201,6 +201,11 @@ class AbstractTitledOrTermedObject(AbstractTitledObject):
         self.sharedTitle = term
         self._process_terms()
 
+    def remove_shared_term(self, term):
+        if self.sharedTitle == term:
+            self.sharedTitle = None
+            self.title_group = self.title_group.copy()
+            return 1
 
 class Term(abst.AbstractMongoRecord, AbstractTitledObject):
     """
@@ -780,6 +785,20 @@ class TitledTreeNode(TreeNode, AbstractTitledOrTermedObject):
 
         if self.sharedTitle and Term().load({"name": self.sharedTitle}).titles != self.get_titles_object():
             raise IndexSchemaError(u"Schema node {} with sharedTitle can not have explicit titles".format(self))
+
+        # disable this check while data is still not conforming to validation
+        if not self.sharedTitle and False:
+            special_book_cases = ["Genesis", "Exodus", "Leviticus", "Numbers", "Deuteronomy", "Judges"]
+            for title in self.title_group.titles:
+                title = title["text"]
+                if self.get_primary_title() in special_book_cases:
+                    break
+                term = Term().load_by_title(title)
+                if term:
+                    if "scheme" in vars(term).keys():
+                        if vars(term)["scheme"] == "Parasha":
+                            raise InputError(
+                                "Nodes that represent Parashot must contain the corresponding sharedTitles.")
 
         #if not self.default and not self.primary_title("he"):
         #    raise IndexSchemaError("Schema node {} missing primary Hebrew title".format(self.key))
@@ -1416,14 +1435,6 @@ class VirtualNode(TitledTreeNode):
         pass
 
 
-class DictionaryEntryNotFound(InputError):
-    def __init__(self, message, lexicon_name=None, base_title=None, word=None):
-        super(DictionaryEntryNotFound, self).__init__(message)
-        self.lexicon_name = lexicon_name
-        self.base_title = base_title
-        self.word = word
-
-
 class DictionaryEntryNode(TitledTreeNode):
     is_virtual = True
     supported_languages = ["en"]
@@ -1476,7 +1487,7 @@ class DictionaryEntryNode(TitledTreeNode):
             self.has_word_match = bool(self.lexicon_entry)
 
         if not self.word or not self.has_word_match:
-            raise DictionaryEntryNotFound("Word not found in {}".format(self.parent.full_title()), self.parent.lexiconName, self.parent.full_title(), self.word)
+            raise DictionaryEntryNotFoundError("Word not found in {}".format(self.parent.full_title()), self.parent.lexiconName, self.parent.full_title(), self.word)
 
     def __eq__(self, other):
         return self.address() == other.address()
@@ -1587,13 +1598,13 @@ class DictionaryNode(VirtualNode):
     def first_child(self):
         try:
             return self.entry_class(self, word=self.firstWord)
-        except DictionaryEntryNotFound:
+        except DictionaryEntryNotFoundError:
             return None
 
     def last_child(self):
         try:
             return self.entry_class(self, word=self.lastWord)
-        except DictionaryEntryNotFound:
+        except DictionaryEntryNotFoundError:
             return None
 
     def all_children(self):
@@ -1660,7 +1671,7 @@ class SheetNode(NumberedTitledTreeNode):
 
         self.sheet_object = db.sheets.find_one({"id": int(self.sheetId)})
         if not self.sheet_object:
-            raise InputError
+            raise SheetNotFoundError
 
     def has_numeric_continuation(self):
         return False  # What about section level?
@@ -2159,6 +2170,6 @@ class AddressSeif(AddressInteger):
 
 class AddressSection(AddressInteger):
     section_patterns = {
-        "en": ur"""(?:(?:([Ss]ections?|§)?\s*)""",  #  the internal ? is a hack to allow a non match, even if 'strict'
+        "en": ur"""(?:(?:([Ss]ections?|§)?\s*))""",  #  the internal ? is a hack to allow a non match, even if 'strict'
         "he": ur""""""
     }
